# VSLab/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "VSLab, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/ChessPlayers": {
      "get": {
        "tags": [
          "ChessPlayers"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/dtoChessPlayer"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ChessPlayers"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/dtoChessPlayer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/tblChessPlayer"
            }
          }
        }
      }
    },
    "/api/ChessPlayers/{id}": {
      "get": {
        "tags": [
          "ChessPlayers"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/tblChessPlayer"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ChessPlayers"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/dtoChessPlayer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ChessPlayers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "dtoChessPlayer": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "rating": {
          "format": "int32",
          "type": "integer"
        },
        "isMaster": {
          "format": "int32",
          "type": "integer"
        },
        "startYear": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "tblChessChampion": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "lastTrophy": {
          "type": "string"
        },
        "record": {
          "type": "string"
        },
        "maxRating": {
          "format": "int32",
          "type": "integer"
        },
        "consecutiveYears": {
          "format": "int32",
          "type": "integer"
        },
        "current": {
          "format": "int32",
          "type": "integer"
        },
        "chessPlayerID": {
          "format": "int32",
          "type": "integer"
        },
        "chessPlayer": {
          "$ref": "#/definitions/tblChessPlayer"
        }
      }
    },
    "tblChessParticipation": {
      "type": "object",
      "properties": {
        "dateSigned": {
          "type": "string"
        },
        "durationPlayed": {
          "type": "string"
        },
        "chessPlayerID": {
          "format": "int32",
          "type": "integer"
        },
        "chessTournamentID": {
          "format": "int32",
          "type": "integer"
        },
        "chessPlayer": {
          "$ref": "#/definitions/tblChessPlayer"
        },
        "chessTournament": {
          "$ref": "#/definitions/tblChessTournament"
        }
      }
    },
    "tblChessPlayer": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "rating": {
          "format": "int32",
          "type": "integer"
        },
        "isMaster": {
          "format": "int32",
          "type": "integer"
        },
        "startYear": {
          "format": "int32",
          "type": "integer"
        },
        "chessChampions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/tblChessChampion"
          }
        },
        "playerParticipations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/tblChessParticipation"
          }
        },
        "chessTournaments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/tblChessTournament"
          }
        }
      }
    },
    "tblChessTournament": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "numParticipants": {
          "format": "int32",
          "type": "integer"
        },
        "host": {
          "type": "string"
        },
        "prizeMoney": {
          "format": "int32",
          "type": "integer"
        },
        "trophy": {
          "type": "string"
        },
        "tournamentParticipations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/tblChessParticipation"
          }
        },
        "chessPlayers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/tblChessPlayer"
          }
        }
      }
    }
  }
}